+++ packages/EasyCodingStandard/composer.json
@@ -20,8 +20,7 @@
 
         "symplify/package-builder": "2.0-RC2 as 1.4",
         "friendsofphp/php-cs-fixer": "^2.3",
-        "squizlabs/php_codesniffer": "dev-master",
-        "contributte/console": "^0.1"
+        "squizlabs/php_codesniffer": "dev-master"
     },
     "require-dev": {
         "phpunit/phpunit": "^6.0",
