+++ packages/EasyCodingStandard/src/DependencyInjection/CompilerPass/CollectorCompilerPass.php
@@ -2,20 +2,36 @@
 
 namespace Symplify\EasyCodingStandard\DependencyInjection\CompilerPass;
 
+use PHP_CodeSniffer\Sniffs\Sniff;
 use PhpCsFixer\Fixer\FixerInterface;
+use Symfony\Component\Console\Application as ConsoleApplication;
+use Symfony\Component\Console\Command\Command;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symplify\EasyCodingStandard\Application\Application;
 use Symplify\EasyCodingStandard\Contract\Application\FileProcessorInterface;
 use Symplify\EasyCodingStandard\FixerRunner\Application\FixerFileProcessor;
+use Symplify\EasyCodingStandard\SniffRunner\Contract\SniffCollectorInterface;
 use Symplify\PackageBuilder\Adapter\Symfony\DependencyInjection\DefinitionCollector;
 
 final class CollectorCompilerPass implements CompilerPassInterface
 {
     public function process(ContainerBuilder $containerBuilder): void
     {
+        $this->collectCommandsToConsoleApplication($containerBuilder);
         $this->collectFileProcessorsToApplication($containerBuilder);
         $this->collectFixersToFixerFileProcessor($containerBuilder);
+        $this->collectSniffsToSniffFileProcessor($containerBuilder);
+    }
+
+    private function collectCommandsToConsoleApplication(ContainerBuilder $containerBuilder): void
+    {
+        DefinitionCollector::loadCollectorWithType(
+            $containerBuilder,
+            ConsoleApplication::class,
+            Command::class,
+            'add'
+        );
     }
 
     private function collectFileProcessorsToApplication(ContainerBuilder $containerBuilder): void
@@ -35,6 +51,16 @@ final class CollectorCompilerPass implements CompilerPassInterface
             FixerFileProcessor::class,
             FixerInterface::class,
             'addFixer'
+        );
+    }
+
+    private function collectSniffsToSniffFileProcessor(ContainerBuilder $containerBuilder): void
+    {
+        DefinitionCollector::loadCollectorWithType(
+            $containerBuilder,
+            SniffCollectorInterface::class,
+            Sniff::class,
+            'addSniff'
         );
     }
 }
